requirements.txt
python-dotenv==1.0.0
cryptography==41.0.7
rich==13.7.0
pydantic==2.6.1
psutil==5.9.8
ttkthemes==3.2.2 